requirements.txt
flask==3.0.3
python-dotenv==1.0.1
openai==1.70.0
python-pptx==1.0.2
gunicorn==22.0.0
Werkzeug==3.0.3
Jinja2==3.1.4 